+++ src/main/java/za/ac/cput/service/ICustomerOrderService.java
@@ -4,6 +4,6 @@ import za.ac.cput.domain.CustomerOrder;
 
 import java.util.List;
 
-public interface ICustomerOrderService extends IService<CustomerOrder, Integer> {
+public interface ICustomerOrderService extends IService<CustomerOrder, Long> {
     List<CustomerOrder> getAll();
 }
